Move zodiac data to .json file
it seems that require() can read and parse json synchronously, with
caching 👍
