--- data/zodiac-signs.js
@@ -1,16 +1,0 @@
-const zodiacSigns = [
-  { id: '1', name: 'aquarius', dates: 'Jan 21 - Feb 19', symbol: 'Water Bearer' },
-  { id: '2', name: 'aries', dates: 'Mar 21 - Apr 19', symbol: 'Ram' },
-  { id: '3', name: 'cancer', dates: 'Jun 21 - Jul 22', symbol: 'Crab' },
-  { id: '4', name: 'capricorn', dates: 'Dec 22 - Feb 19', symbol: 'Goat' },
-  { id: '5', name: 'gemini', dates: 'May 21 - Jun 20', symbol: 'Twins' },
-  { id: '6', name: 'leo', dates: 'Jul 23 - Aug 22', symbol: 'Lion' },
-  { id: '7', name: 'libra', dates: 'Sept 23 - Oct 22', symbol: 'Balance' },
-  { id: '8', name: 'pisces', dates: 'Feb 19 - Mar 20', symbol: 'Fish' },
-  { id: '9', name: 'sagittarius', dates: 'Nov 22 - Dec 21', symbol: 'Archer' },
-  { id: '10', name: 'scorpio', dates: 'Oct 23 - Nov 21', symbol: 'Scorpion' },
-  { id: '11', name: 'taurus', dates: 'Apr 20 - May 20', symbol: 'Bull' },
-  { id: '12', name: 'virgo', dates: 'Aug 23 - Sep 22', symbol: 'Virgin' },
-];
-
-module.exports = { zodiacSigns };
