+++ lib/controllers/zodiac.js
@@ -1,5 +1,5 @@
 const { Router } = require('express');
-const { zodiacSigns } = require('../../data/zodiac-signs');
+const zodiacSigns = require('../../data/zodiac-signs.json');
 
 const router = new Router();
